changed name parameter for user to fullName, to avoid confusion with username (which will have a unique check when setting up to be used as PK)
