--- Client/src/View/BookInforViewController.java
@@ -1,5 +1,0 @@
-package View;
-
-public class BookInforViewController
-{
-}
