--- Client/src/View/BookListViewController.java
@@ -1,5 +1,0 @@
-package View;
-
-public class BookListViewController
-{
-}
