--- Client/src/View/NotificationViewController.java
@@ -1,5 +1,0 @@
-package View;
-
-public class NotificationViewController
-{
-}
