--- Client/src/ViewModel/BookListVM.java
@@ -1,5 +1,0 @@
-package ViewModel;
-
-public class BookListVM
-{
-}
