--- Client/src/ViewModel/LogInVM.java
@@ -1,6 +1,0 @@
-package ViewModel;
-
-public class LogInVM
-{
-
-}
