--- Client/src/ViewModel/NotificationsVM.java
@@ -1,5 +1,0 @@
-package ViewModel;
-
-public class NotificationsVM
-{
-}
