--- Client/src/ViewModel/ReaderNotesVM.java
@@ -1,5 +1,0 @@
-package ViewModel;
-
-public class ReaderNotesVM
-{
-}
