--- Client/src/ViewModel/UserProfileVM.java
@@ -1,5 +1,0 @@
-package ViewModel;
-
-public class UserProfileVM
-{
-}
