+++ Client/src/view/BookInforViewController.java
@@ -1,0 +1,5 @@
+package view;
+
+public class BookInforViewController
+{
+}
