+++ Client/src/view/BookListViewController.java
@@ -1,0 +1,5 @@
+package view;
+
+public class BookListViewController
+{
+}
