+++ Client/src/view/LoginView.fxml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.image.*?>
+<?import javafx.scene.layout.*?>
+
+<TitledPane fx:controller="view.LoginController" animated="false" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" text="User Log in" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1">
+  <content>
+    <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="278.0" prefWidth="598.0">
+         <children>
+            <ImageView id="logo" fitHeight="150.0" fitWidth="200.0" layoutX="224.0" layoutY="35.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../../../img/logo_mockup.png" />
+               </image>
+            </ImageView>
+            <TextField id="usernameField" layoutX="123.0" layoutY="208.0" prefHeight="26.0" prefWidth="351.0" promptText="username" />
+            <PasswordField id="passwordField" layoutX="123.0" layoutY="258.0" prefHeight="26.0" prefWidth="352.0" promptText="password" />
+            <Button id="loginButton"  onAction="#onLoginClicked" layoutX="216.0" layoutY="313.0" mnemonicParsing="false" text="Log in" />
+            <Button id="cancelButton" onAction = "#onCancelClicked" layoutX="330.0" layoutY="313.0" mnemonicParsing="false" text="Cancel" />
+         </children></AnchorPane>
+  </content>
+</TitledPane>
