+++ Client/src/view/NotificationViewController.java
@@ -1,0 +1,5 @@
+package view;
+
+public class NotificationViewController
+{
+}
