+++ Client/src/viewmodel/BookListVM.java
@@ -1,0 +1,5 @@
+package viewmodel;
+
+public class BookListVM
+{
+}
