+++ Client/src/viewmodel/LogInVM.java
@@ -1,0 +1,6 @@
+package viewmodel;
+
+public class LogInVM
+{
+
+}
