+++ Client/src/viewmodel/NotificationsVM.java
@@ -1,0 +1,5 @@
+package viewmodel;
+
+public class NotificationsVM
+{
+}
