+++ Client/src/viewmodel/ReaderNotesVM.java
@@ -1,0 +1,5 @@
+package viewmodel;
+
+public class ReaderNotesVM
+{
+}
