+++ Client/src/viewmodel/UserProfileVM.java
@@ -1,0 +1,5 @@
+package viewmodel;
+
+public class UserProfileVM
+{
+}
