+++ Server/ServerTest/DAOtesting/InMemoryUserDAO.java
@@ -1,0 +1,29 @@
+package DAOtesting;
+
+import model.User;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+
+public class InMemoryUserDAO implements UserDAO
+{
+  private final Map<String, User> users = new ConcurrentHashMap<>();
+
+  @Override
+  public Optional<User> findByUserName(String username)
+  {
+    return Optional.ofNullable(users.get(username));
+  }
+
+  @Override
+  public void save(User u)
+  {
+    if (users.containsKey(u.getUserName()))
+    {
+      throw new IllegalArgumentException("Username already taken");
+    }
+    users.put(u.getUserName(), u);
+  }
+}
+
