+++ Server/ServerTest/TestUser.java
@@ -1,5 +1,5 @@
 
-import Model.User;
+import model.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -137,7 +137,7 @@ public class TestUser
   @Test
   public void testGetPassword()
   {
-    assertEquals("1234", testUser.getPassword());
+    assertEquals("1234", testUser.getPasswordHash());
   }
 
   //phone number tests
