--- Server/src/Model/ActionManagers/MarkingAsRead.java
@@ -1,5 +1,0 @@
-package Model.ActionManagers;
-
-public class MarkingAsRead
-{
-}
