--- Server/src/Model/Library.java
@@ -1,14 +1,0 @@
-package Model;
-
-import java.util.List;
-
-public class Library
-{
-  private List<Book> books;
-
-  public Library(List<Book> books)
-  {
-    this.books = books;
-  }
-
-}
