--- Server/src/Model/Status/Available.java
@@ -1,5 +1,0 @@
-package Model.Status;
-
-public class Available implements Status
-{
-}
