--- Server/src/Model/Status/Borrowed.java
@@ -1,5 +1,0 @@
-package Model.Status;
-
-public class Borrowed implements Status
-{
-}
