--- Server/src/Model/Status/Status.java
@@ -1,5 +1,0 @@
-package Model.Status;
-
-public interface Status
-{
-}
