--- Server/src/Model/Status/Unavailable.java
@@ -1,5 +1,0 @@
-package Model.Status;
-
-public class Unavailable implements Status
-{
-}
