+++ Server/src/model/Library.java
@@ -1,0 +1,14 @@
+package model;
+
+import java.util.List;
+
+public class Library
+{
+  private List<Book> books;
+
+  public Library(List<Book> books)
+  {
+    this.books = books;
+  }
+
+}
