+++ Server/src/model/actionmanagers/MarkingAsRead.java
@@ -1,0 +1,5 @@
+package model.actionmanagers;
+
+public class MarkingAsRead
+{
+}
