+++ Server/src/model/status/Available.java
@@ -1,0 +1,5 @@
+package model.status;
+
+public class Available implements Status
+{
+}
