+++ Server/src/model/status/Borrowed.java
@@ -1,0 +1,5 @@
+package model.status;
+
+public class Borrowed implements Status
+{
+}
