+++ Server/src/model/status/Status.java
@@ -1,0 +1,5 @@
+package model.status;
+
+public interface Status
+{
+}
