+++ Server/src/model/status/Unavailable.java
@@ -1,0 +1,5 @@
+package model.status;
+
+public class Unavailable implements Status
+{
+}
