--- Shared/src/DTO/BookDTO.java
@@ -1,5 +1,0 @@
-package DTO;
-
-public class BookDTO
-{
-}
