--- Shared/src/DTO/ReaderNoteDTO.java
@@ -1,5 +1,0 @@
-package DTO;
-
-public class ReaderNoteDTO
-{
-}
