--- Shared/src/DTO/Request.java
@@ -1,5 +1,0 @@
-package DTO;
-
-public class Request
-{
-}
