--- Shared/src/DTO/Response.java
@@ -1,5 +1,0 @@
-package DTO;
-
-public class Response
-{
-}
