--- Shared/src/DTO/UserDTO.java
@@ -1,5 +1,0 @@
-package DTO;
-
-public class UserDTO
-{
-}
