+++ Shared/src/dto/BookDTO.java
@@ -1,0 +1,5 @@
+package dto;
+
+public class BookDTO
+{
+}
