+++ Shared/src/dto/ReaderNoteDTO.java
@@ -1,0 +1,5 @@
+package dto;
+
+public class ReaderNoteDTO
+{
+}
