+++ Shared/src/dto/Request.java
@@ -1,0 +1,5 @@
+package dto;
+
+public class Request
+{
+}
