+++ Shared/src/dto/Response.java
@@ -1,0 +1,5 @@
+package dto;
+
+public class Response
+{
+}
