+++ Shared/src/dto/UserDTO.java
@@ -1,0 +1,5 @@
+package dto;
+
+public class UserDTO
+{
+}
